refactor(file-tree): consolidate listDirectory + gitignore under file/utils
Move `search.ts` (ripgrep + fuzzy + `listDirectory`) and `gitignore.ts` (the
.gitignore predicate loader) out of `@main/utils/file/` and `file/tree/`
into `@main/file/utils/`, so the `file/` namespace owns every primitive the
directory tree relies on without crossing back into the legacy
`utils/file/` surface.

- Tree builder + FileManager IPC handler now import from `@main/file/utils/search`.
- `gitignore.ts` was tree-internal but conceptually a general FS helper;
  moving it next to search keeps watcher-ignore wiring + tree builder
  fed from one place without a tree dependency.
- New `@main/file/utils/index.ts` is doc-only — same subpath-import-only
  convention as `@main/utils/file/index.ts` to avoid barrel cycles.
- `@main/utils/file/index.ts` doc updated to point at the new location.

(cherry picked from commit ad72f0e12a692c3d5c58250c066aeb44abe1de77)
