+++ src/main/file/utils/gitignore.ts
