+++ src/main/utils/file/index.ts
@@ -8,19 +8,22 @@
  *   `pathExists`, `directoryExists`, `isPathInside`, `untildify`,
  *   `hasWritePermission`, `resolveAndValidatePath`, …). Re-exported through
  *   this barrel; callers import with `from '@main/utils/file'`.
- * - `./{fs,metadata,path,search,shell}` — pure FS primitives. Access via
+ * - `./{fs,metadata,path,shell}` — pure FS primitives. Access via
  *   **explicit subpath imports**, e.g.
  *   `import { atomicWriteFile } from '@main/utils/file/fs'`. Not re-exported
  *   through the barrel to avoid symbol collisions with legacy helpers
  *   (notably `getFileType`, which has different signatures in the two
  *   modules: legacy takes an extension, primitive takes a path).
+ * - Directory-listing (`listDirectory`) and `.gitignore` parsing live next
+ *   to the consumer that owns them: `@main/file/utils/{search,gitignore}`.
+ *   See `src/main/file/utils/` for that surface.
  *
  * These modules will consolidate over time — `legacyFile.ts` and the sibling
  * `../fileOperations.ts` are expected to be split into the primitive modules
  * above (fs/metadata/path/…), after which this barrel can expose the
  * primitive surface directly.
  *
- * ## Access policy for the FS primitives (fs/metadata/path/search/shell)
+ * ## Access policy for the FS primitives (fs/metadata/path/shell)
  *
  * These are the **sole FS owners** for the main process — callers like
  * `BootConfigService`, the MCP OAuth flow, and any service that truly needs
